requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0
pydantic==2.10.4
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.33.0
torch==1.9.0
torchvision==0.10.0